+++ src/enums/serviceAliases.ts
@@ -71,5 +71,6 @@ export default {
   [services.systemsManagerDocument]: 'systemsManagerDocuments',
   [services.systemsManagerInstance]: 'systemsManagerInstances',
   [services.transitGatewayAttachment]: 'transitGatewayAttachments',
+  [services.vpcEndpoint]: 'vpcEndpoints',
   [services.vpnConnection]: 'vpnConnections',
 }